fix(restructure): restructured the for components repeated functionality
