+++ .prettierrc
@@ -1,12 +1,12 @@
 {
-    "semi": true,
-    "trailingComma": "es5",
-    "singleQuote": true,
-    "tabWidth": 2,
-    "printWidth": 80,
-    "jsxSingleQuote": false,
-    "bracketSpacing": true,
-    "arrowParens": "always",
-    "endOfLine": "auto"
-  }
-  
+  "semi": true,
+  "trailingComma": "es5",
+  "singleQuote": true,
+  "tabWidth": 2,
+  "printWidth": 80,
+  "jsxSingleQuote": true,
+  "bracketSpacing": true,
+  "arrowParens": "always",
+  "endOfLine": "lf",
+  "bracketSameLine": true
+}
