+++ index.html
@@ -5,6 +5,11 @@
     <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Sticky-Memo</title>
+    <meta name="description" content="Sticky-Memo is a web application for creating and managing sticky notes." />
+    <meta name="keywords" content="Sticky, Memo, Notes, Web Application" />
+    <meta name="author" content="Narain Karthik" />
+    <link href="https://fonts.googleapis.com/css2?family=Outfit&display=swap" rel="stylesheet">
+    <link rel="preload" href="/src/main.jsx" as="script" />
   </head>
   <body>
     <div id="root"></div>
