+++ src/App.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Navbar from "./components/Navbar";
+import Navbar from "./components/common/Navbar";
 import NoteList from "./pages/NoteList";
 import TableList from "./pages/TableList";
 import BoardList from "./pages/BoardList";
