+++ src/components/Board/BoardCard.jsx
@@ -6,9 +6,10 @@ import CheckCircleIcon from '@mui/icons-material/CheckCircle';
 import DeleteIcon from '@mui/icons-material/Delete';
 import EditIcon from '@mui/icons-material/Edit';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
-import { holdItem, checkItem, deleteItem } from '../utils/helper';
+import { holdItem, checkItem, deleteItem } from '../../utils/helper';
+import { cardStyles, buttonStyle, typographyStyles, popoverStyles, textFieldStyles } from './styles';
 
-const NoteCard = ({
+const BoardCard = ({
   item,
   index,
   isEditing,
@@ -31,12 +32,6 @@ const NoteCard = ({
 }) => {
   const open = Boolean(anchorEl);
   const id = open ? 'simple-popover' : undefined;
-  const buttonStyle = {
-    color: 'white',
-    borderRadius: '100%',
-    backgroundColor: '#f66b6b',
-    padding: '3px',
-  };
 
   return (
     <Card
@@ -45,35 +40,13 @@ const NoteCard = ({
       onDragStart={() => handleDragStart(index)}
       onDrop={(e) => handleDrop(index, e)}
       onDragOver={(e) => handleDragOver(e)}
-      sx={{
-        display: 'flex',
-        flexDirection: 'column',
-        backgroundColor: item.checked ? 'rgb(119, 237, 119)' : item.held ? 'rgb(115, 191, 238)' : 'rgb(248, 248, 154)',
-        borderRadius: '8px',
-        width: '250px',
-        height: '175px',
-        padding: 0,
-        transition: 'transform 0.3s ease-in, background-color 0.3s ease-in',
-        margin: '20px 15px 0px 0px',
-        border: 'none',
-        overflow: 'hidden',
-        '&:hover': { transform: 'scale(1.01)' },
-      }}
+      sx={cardStyles(item)}
     >
       <CardContent sx={{ padding: 0 }}>
-        <Typography
-          sx={{
-            backgroundColor: '#3498DB',
-            height: '2.6em',
-            color: 'white',
-            fontWeight: 'bold',
-            alignContent: 'center',
-            position: 'relative'
-          }}
-        >
+        <Typography sx={typographyStyles}>
           <>
             {isEditing && editingIndex === index ? (
-              <TextField size='small' onChange={(e) => setEditedTitle(e.target.value)} defaultValue={item.title} ></TextField>
+              <TextField size='small' onChange={(e) => setEditedTitle(e.target.value)} defaultValue={item.title} fullWidth />
             ) :
               (
                 <span style={{ fontWeight: 'bolder' }}>{item.title}</span>
@@ -82,7 +55,7 @@ const NoteCard = ({
           </>
           <>
             <IconButton
-              sx={{ position: 'absolute', top: 0, right: 0, backgroundColor: '#f66b6b', padding: '3px', margin: '8px', color: 'white' }}
+              sx={buttonStyle}
               aria-describedby={id}
               onClick={(e) => handleClickPopover(e, index)}
             >
@@ -99,13 +72,7 @@ const NoteCard = ({
                 horizontal: 'right',
               }}
             >
-              <Typography
-                sx={{
-                  backgroundColor: '#3498db',
-                  padding: '0.5',
-                  alignSelf: 'end'
-                }}
-              >
+              <Typography sx={popoverStyles}>
                 <IconButton
                   onClick={() => {
                     if (isEditing) {
@@ -122,27 +89,29 @@ const NoteCard = ({
                     <EditIcon fontSize="small" sx={buttonStyle} />
                   )}
                 </IconButton>
-                <IconButton onClick={() => holdItem(setItems, editingIndex, setSnackbar, 'Note')} variant="contained">
+                <IconButton onClick={() => holdItem(setItems, editingIndex, setSnackbar, 'Board')} variant="contained">
                   <BackHandIcon fontSize="small" sx={buttonStyle} />
                 </IconButton>
-                <IconButton onClick={() => checkItem(setItems, editingIndex, setSnackbar, 'Note')} variant="contained">
+                <IconButton onClick={() => checkItem(setItems, editingIndex, setSnackbar, 'Board')} variant="contained">
                   <CheckCircleIcon fontSize="small" sx={buttonStyle} />
                 </IconButton>
-                <IconButton onClick={() => deleteItem(setItems, editingIndex, setSnackbar, 'Note')} variant="contained">
+                <IconButton onClick={() => deleteItem(setItems, editingIndex, setSnackbar, 'Board')} variant="contained">
                   <DeleteIcon fontSize="small" sx={buttonStyle} />
                 </IconButton>
               </Typography>
             </Popover>
           </>
         </Typography>
-        <Typography>
+        <Typography sx={textFieldStyles}>
           {isEditing && editingIndex === index ? (
             <TextField
               size='small'
               sx={{ width: '100%' }}
               onChange={(e) => setEditedContent(e.target.value)}
               defaultValue={item.content}
-            ></TextField>
+              multiline
+              rows={4}
+            />
           ) :
             (
               <span>{item.content}</span>
@@ -154,4 +123,4 @@ const NoteCard = ({
   );
 };
 
-export default NoteCard;
+export default BoardCard;
