+++ src/components/Note/NoteCard.jsx
@@ -6,9 +6,10 @@ import CheckCircleIcon from '@mui/icons-material/CheckCircle';
 import DeleteIcon from '@mui/icons-material/Delete';
 import EditIcon from '@mui/icons-material/Edit';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
-import { holdItem, checkItem, deleteItem } from '../utils/helper';
+import { holdItem, checkItem, deleteItem } from '../../utils/helper';
+import { cardStyles, buttonStyle, typographyStyles, popoverStyles, textFieldStyles } from './styles';
 
-const BoardCard = ({
+const NoteCard = ({
   item,
   index,
   isEditing,
@@ -31,12 +32,6 @@ const BoardCard = ({
 }) => {
   const open = Boolean(anchorEl);
   const id = open ? 'simple-popover' : undefined;
-  const buttonStyle = {
-    color: 'white',
-    borderRadius: '100%',
-    backgroundColor: '#f66b6b',
-    padding: '3px',
-  };
 
   return (
     <Card
@@ -45,32 +40,10 @@ const BoardCard = ({
       onDragStart={() => handleDragStart(index)}
       onDrop={(e) => handleDrop(index, e)}
       onDragOver={(e) => handleDragOver(e)}
-      sx={{
-        display: 'flex',
-        flexDirection: 'column',
-        backgroundColor: item.checked ? 'rgb(119, 237, 119)' : item.held ? 'rgb(115, 191, 238)' : 'rgb(248, 248, 154)',
-        borderRadius: '12px',
-        width: '300px',
-        height: '200px',
-        padding: '1em',
-        transition: 'transform 0.3s ease-in, background-color 0.3s ease-in',
-        margin: '20px 15px 0px 0px',
-        boxShadow: '0 4px 8px rgba(0, 0, 0, 0.1)',
-        '&:hover': { transform: 'scale(1.02)', boxShadow: '0 6px 12px rgba(0, 0, 0, 0.15)' },
-      }}
+      sx={cardStyles(item)}
     >
       <CardContent sx={{ padding: 0 }}>
-        <Typography
-          sx={{
-            height: '2em',
-            fontWeight: 'bold',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'space-between',
-            padding: '0 1em',
-            borderRadius: '8px 8px 0 0',
-          }}
-        >
+        <Typography sx={typographyStyles}>
           <>
             {isEditing && editingIndex === index ? (
               <TextField size='small' onChange={(e) => setEditedTitle(e.target.value)} defaultValue={item.title} fullWidth />
@@ -82,7 +55,7 @@ const BoardCard = ({
           </>
           <>
             <IconButton
-              sx={{ backgroundColor: '#f66b6b', padding: '3px', color: 'white' }}
+              sx={buttonStyle}
               aria-describedby={id}
               onClick={(e) => handleClickPopover(e, index)}
             >
@@ -99,14 +72,7 @@ const BoardCard = ({
                 horizontal: 'right',
               }}
             >
-              <Typography
-                sx={{
-                  backgroundColor: '#3498db',
-                  padding: '0.5',
-                  display: 'flex',
-                  justifyContent: 'space-around',
-                }}
-              >
+              <Typography sx={popoverStyles}>
                 <IconButton
                   onClick={() => {
                     if (isEditing) {
@@ -123,20 +89,20 @@ const BoardCard = ({
                     <EditIcon fontSize="small" sx={buttonStyle} />
                   )}
                 </IconButton>
-                <IconButton onClick={() => holdItem(setItems, editingIndex, setSnackbar, 'Board')} variant="contained">
+                <IconButton onClick={() => holdItem(setItems, editingIndex, setSnackbar, 'Note')} variant="contained">
                   <BackHandIcon fontSize="small" sx={buttonStyle} />
                 </IconButton>
-                <IconButton onClick={() => checkItem(setItems, editingIndex, setSnackbar, 'Board')} variant="contained">
+                <IconButton onClick={() => checkItem(setItems, editingIndex, setSnackbar, 'Note')} variant="contained">
                   <CheckCircleIcon fontSize="small" sx={buttonStyle} />
                 </IconButton>
-                <IconButton onClick={() => deleteItem(setItems, editingIndex, setSnackbar, 'Board')} variant="contained">
+                <IconButton onClick={() => deleteItem(setItems, editingIndex, setSnackbar, 'Note')} variant="contained">
                   <DeleteIcon fontSize="small" sx={buttonStyle} />
                 </IconButton>
               </Typography>
             </Popover>
           </>
         </Typography>
-        <Typography sx={{ padding: '1em', flexGrow: 1 }}>
+        <Typography sx={textFieldStyles}>
           {isEditing && editingIndex === index ? (
             <TextField
               size='small'
@@ -157,4 +123,4 @@ const BoardCard = ({
   );
 };
 
-export default BoardCard;
+export default NoteCard;
