+++ src/components/Table/TableCard.jsx
@@ -6,7 +6,8 @@ import CheckCircleIcon from '@mui/icons-material/CheckCircle';
 import DeleteIcon from '@mui/icons-material/Delete';
 import EditIcon from '@mui/icons-material/Edit';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
-import { holdItem, checkItem, deleteItem } from '../utils/helper';
+import { holdItem, checkItem, deleteItem } from '../../utils/helper';
+import { tableRowStyles, buttonStyle, boxStyles, contentBoxStyles, popoverTypographyStyles } from './styles'; // Import specific styles
 
 const TableCard = ({
   item,
@@ -31,12 +32,6 @@ const TableCard = ({
 }) => {
   const open = Boolean(anchorEl);
   const id = open ? 'simple-popover' : undefined;
-  const buttonStyle = {
-    color: 'white',
-    borderRadius: '100%',
-    backgroundColor: '#f66b6b',
-    padding: '3px',
-  };
 
   return (
     <TableRow
@@ -46,16 +41,7 @@ const TableCard = ({
       onDragStart={(e) => handleDragStart(index)}
       onDrop={(e) => handleDrop(index, e)}
       onDragOver={(e) => handleDragOver(e)}
-      sx={{
-        color: 'black',
-        backgroundColor: item.checked ? 'rgb(119, 237, 119)' : item.held ? 'rgb(115, 191, 238)' : 'rgb(248, 248, 154)',
-        width: '100%',
-        height: 'auto',
-        position: 'relative',
-        zIndex: 1,
-        transition: 'transform 0.3s ease-in, background-color 0.3s ease-in',
-        '&:hover': { transform: 'scale(1.01)' },
-      }}
+      sx={tableRowStyles(item)} // Apply the styles
     >
       <TableCell align='center'>
         {isEditing && editingIndex === index ? (
@@ -65,8 +51,8 @@ const TableCard = ({
         )}
       </TableCell>
       <TableCell align='center'>
-        <Box style={{ display: 'flex', justifyContent: 'space-between' }}>
-          <Box style={{ alignSelf: 'center', wordBreak: 'break-all' }}>
+        <Box sx={boxStyles}>
+          <Box sx={contentBoxStyles}>
             {isEditing && editingIndex === index ? (
               <TextField onChange={(e) => setEditedContent(e.target.value)} defaultValue={item.content} fullWidth multiline rows={4} />
             ) : (
@@ -83,7 +69,7 @@ const TableCard = ({
             onClose={handleClosePopover}
             anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
           >
-            <Typography sx={{ backgroundColor: '#3498db', padding: '0.5', display: 'flex', justifyContent: 'space-around' }}>
+            <Typography sx={popoverTypographyStyles}>
               <IconButton
                 onClick={() => {
                   if (isEditing) {
