+++ src/pages/BoardList.jsx
@@ -1,14 +1,15 @@
-import React, { useState } from 'react';
+import React, { useState, useTransition } from 'react';
 import { useRecoilState } from 'recoil';
 import { itemsState, snackbarState } from '../utils/state';
-import CreateBoard from '../components/CreateBoard';
-import Footer from "../components/Footer";
-import BoardCard from '../components/BoardCard';
-import CommonFilter from '../components/CommonFilter';
-import CommonSnackbar from '../components/CommonSnackbar';
+import CreateBoard from '../components/Board/CreateBoard';
+import Footer from "../components/common/Footer";
+import BoardCard from '../components/Board/BoardCard';
+import CommonFilter from '../components/common/CommonFilter';
+import CommonSnackbar from '../components/common/CommonSnackbar';
 import { Box } from '@mui/material';
 import { addItem, filterItems } from '../utils/helper';
 import { useItemUtils } from '../utils/useItemUtils';
+import { boardListStyles, scrollBoxStyles } from '../styles/boardListStyles';
 
 const BoardList = (props) => {
   const [items, setItems] = useRecoilState(itemsState);
@@ -17,6 +18,7 @@ const BoardList = (props) => {
   const [draggingIndex, setDraggingIndex] = useState(null);
   const [editingIndex, setEditingIndex] = useState(null);
   const [anchorEl, setAnchorEl] = useState(null);
+  const [isPending, startTransition] = useTransition();
 
   const {
     isEditing,
@@ -36,7 +38,7 @@ const BoardList = (props) => {
       const updatedItems = [...items];
       const [draggedItem] = updatedItems.splice(draggingIndex, 1);
       updatedItems.splice(index, 0, draggedItem);
-      setItems(updatedItems);
+      startTransition(() => setItems(updatedItems));
     }
     setDraggingIndex(null);
   };
@@ -46,7 +48,7 @@ const BoardList = (props) => {
     const updatedItems = items.map((item, index) =>
       index === id ? { ...item, title: newTitle || item.title, content: newContent || item.content } : item
     );
-    setItems(updatedItems);
+    startTransition(() => setItems(updatedItems));
   };
 
   const filteredItems = filterItems(items, filter);
@@ -57,27 +59,13 @@ const BoardList = (props) => {
   const handleClosePopover = () => setAnchorEl(null);
 
   return (
-    <Box sx={{ display: 'flex', flexDirection: 'column', alignItems: 'center', backgroundColor: '#f4f5f7', gap: '10px' }}>
+    <Box sx={boardListStyles}>
       <CommonSnackbar snackbar={snackbar} setSnackbar={setSnackbar} />
       <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
         <CreateBoard onAdd={(newItem) => addItem(setItems, newItem, setSnackbar, "Board")} />
       </Box>
       <CommonFilter filter={filter} setFilter={setFilter} />
-      <Box
-        sx={{
-          display: 'flex',
-          flexWrap: 'wrap',
-          justifyContent: 'center',
-          padding: '1em',
-          overflowY: 'auto',
-          height: '300px',
-          wordWrap: 'break-word',
-          '&::-webkit-scrollbar': { width: '8px', backgroundColor: '#f1f1f1', borderRadius: '8px' },
-          '&::-webkit-scrollbar-thumb': { backgroundColor: 'lightseagreen', borderRadius: '8px', transition: 'background-color 0.3s ease-in-out' },
-          '&::-webkit-scrollbar-thumb:hover': { backgroundColor: 'lightgreen' },
-          '&::-webkit-scrollbar-track': { backgroundColor: '#f1f1f1', borderRadius: '8px' },
-        }}
-      >
+      <Box sx={scrollBoxStyles}>
         {filteredItems.map((item, index) => (
           <BoardCard
             key={index}
