+++ src/utils/state.js
@@ -1,15 +1,16 @@
 import { atom } from 'recoil';
 import { recoilPersist } from 'recoil-persist';
 
-const { persistAtom } = recoilPersist()
+const { persistAtom } = recoilPersist();
 
+// State to manage items
 export const itemsState = atom({
   key: 'itemsState',
   default: [],
   effects_UNSTABLE: [persistAtom],
 });
 
-// Snackbar state to manage notifications
+// State to manage snackbar notifications
 export const snackbarState = atom({
   key: 'snackbarState',
   default: {
