+++ src/utils/useItemUtils.js
@@ -13,10 +13,12 @@ export const useItemUtils = (props) => {
     severity: "success", // Can be "success", "error", "info", or "warning"
   });
 
+  // Handle closing of the snackbar
   const handleSnackbarClose = () => {
     setSnackbar({ ...snackbar, open: false });
   };
 
+  // Show notification with a message and severity
   const showNotification = (message, severity) => {
     setSnackbar({
       open: true,
@@ -25,33 +27,39 @@ export const useItemUtils = (props) => {
     });
   };
 
+  // Handle delete action
   const handleDelete = () => {
     props.onDelete(props.id);
     showNotification(`${props.type} Deleted`, "error");
   };
 
+  // Handle check action
   const handleCheck = () => {
     props.onCheck(props.id);
     showNotification(`${props.type} Checked`, "success");
   };
 
+  // Handle hold action
   const handleHold = () => {
     props.onHold(props.id);
     showNotification(`${props.type} Held`, "info");
   };
 
+  // Handle edit action
   const handleEdit = () => {
     setEditedContent(null);
     setEditedTitle(null);
     setIsEditing(true);
   };
 
+  // Handle save action
   const handleSave = () => {
     props.onSave(props.id, editedTitle, editedContent);
     setIsEditing(false);
     showNotification(`${props.type} Saved Successfully`, "success");
   };
 
+  // Toggle visibility of buttons
   const toggleButtons = () => {
     setShowButtons((prevState) => !prevState);
   };
